refactor(server): peel ingest+WS routes into routers/ingest.py — server.py is now a pure app shell (router-split, round-5 #51 / R5-25 PR27, FINAL)
Fifteenth and FINAL router peel. Moves the whole ingest family into a leaf
routers/ingest.py: GET /api/ingest/engines, POST /api/ingest/scan, POST /api/ingest,
POST /api/media/{id}/reingest, the @router.websocket("/ws/ingest") progress socket,
and POST /api/ingest/ws — with IngestRequest/ScanRequest, _build_scan_manifest,
_ws_authorized (CSWSH/origin auth), _run_ingest_with_ws, the _ingest_ws_tasks set +
_on_ingest_ws_done GC-prevention callback, and the scan ext buckets.

The shared H3 single-flight slot (_acquire/_release_ingest_slot) and the
ingest_ws IngestBroadcaster are imported from state — verified to be the ONE shared
instance (identity-asserted in the boundary test), so all ingest entrypoints
(REST/reingest/WS) still serialize and can't double-load whisper → OOM. _ALLOWED_ORIGINS
from webguard (same instance); config.BASE_DIR replaces server.ROOT for ingest.py.

server.py 911 → 487: it is now ONLY the app shell — /thumbnails/{name}, the SPA
"/" route, lifespan, CORS/middleware, the include_router wiring, ROOT, and the leaf
re-export blocks. R5-25 (#51) is complete: the ~4531-line monolith is split into 15
routers + 6 leaf service modules (pathres/webguard/export_builders/reqopts/
mediarecords/state), all behaviour-preserving.

Behaviour-preserving: paths/methods/scopes/WS auth unchanged (SPA + CLI unaffected).
4 tests repointed to routers.ingest (IngestRequest / reingest _resolve_media_path
patch / scan ext buckets / reqopts leaf-ownership).
